# doc/EZshopBackEndApplication_swagger.yaml
openapi: 3.0.3
info:
  title: EZShop Back End Application API
  version: 1.0.0
  description: |
    OpenAPI definition for the EZShop backend server application.

    **EZShop Description**

    EZShop is a retail management system designed to handle the end-to-end operations of a small to medium-sized shop, integrating user management, inventory control, sales, orders, and accounting functions into a single platform.
    EZShop defines a set of business operations that orchestrates user interactions, product lifecycle, and financial transactions.

    In a Python-based reimplementation using FastAPI, EZShop would act as a RESTful backend service, exposing endpoints to manage authentication, product catalog, sales transactions, customer information, and payments.
    Authentication would rely on JWT-based bearer tokens, with access permissions differentiated by user roles (Administrator, ShopManager, Cashier).
    
    **Hierarchical Structure and Components**

    The system architecture is logically divided into four layers:

    1. Users and Authentication Layer
       - Manages user accounts and roles (Administrator, Cashier, ShopManager).
       - Provides endpoints for login, logout, and user CRUD operations.
       - Handles session security and authorization for all operations.

    2. Inventory and Product Management Layer
       - Handles product types, quantities, and barcodes (with optional RFID tracking).
...

...

...
      
...
    
...

...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - name
      properties:
        code: 
          type: integer
          description: HTTP status code of the error
        message:
          type: string
          description: message of the error
        name:
          type: string
          description: Descriptive error message
      description: Standard structure for error responses
...
    BooleanResponse:
      type: object
      properties:
        success: 
          type: boolean
      required: [success]
...
    ReturnStatus:
      type: string
      enum: [OPEN, CLOSED, REIMBURSED]
      description: Status of a return transaction
    ReturnLine:
      type: object
      required:
        - return_id
        - product_barcode
        - quantity
        - price_per_unit
      properties:
        id:
          type: integer
          example: 1
        return_id:
          type: integer
          example: 5
        product_barcode:
          type: string
          example: "0123456789012"
        quantity:
          type: integer
          example: 2
        price_per_unit:
          type: number
          format: float
          example: 9.99
    Return:
      type: object
      required:
        - sale_id
        - status
      properties:
        id:
          type: integer
          example: 3
        sale_id:
          type: integer
          example: 1
        status:
          $ref: '#/components/schemas/ReturnStatus'
        created_at:
          type: string
          format: date-time
          example: "2025-11-03T12:30:00Z"
        closed_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-11-04T09:45:00Z"
        lines:
          type: array
          items:
            $ref: '#/components/schemas/ReturnLine'
paths:
...
  /returns:
    post:
      tags:
      - returns
      summary: Start a return transaction for a sale
      description: | 
        Creates a new return transaction for a given **PAID** sale. 
        Requires **Administrator**, **ShopManager**, or **Cashier** role.
      operationId: startReturn
      security:
        - bearerAuth: []
      parameters:
        - name: sale_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Return transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Return'
        '400':
          description: Invalid sale ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "invalid id"
                name: "BadRequestError"
        '404':
          description: Return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Return not found"
                name: "NotFoundError"
        '420':
          description: Invalid sale state for return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 420
                message: "Return allowed only on paid sales"
                name: "InvalidStateError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
    get:
      tags:
        - returns
      summary: Get all return transactions
      description: Retrieves all return transactions. Requires **Administrator**, **ShopManager**, or **Cashier** role.
      operationId: getAllReturns
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all return transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Return'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
  /returns/{return_id}:
    get:
      tags:
        - returns
      summary: Get a return transaction by ID
      description: Retrieves details of a specific return transaction. Requires **Administrator**, **ShopManager**, or **Cashier** role.
      operationId: getReturn
      security:
        - bearerAuth: []
      parameters:
        - name: return_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Return transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Return'
        '400':
          description: Invalid or missing id. 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Bad request"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '404':
          description: Return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Return not found"
                name: "NotFoundError"
    delete:
      tags:
        - returns
      summary: Delete a return transaction
      description: | 
        Deletes a return transaction if not already **REIMBURSED**. 
        Requires **Administrator**, **ShopManager**, or **Cashier** role.
      operationId: deleteReturn
      security:
        - bearerAuth: []
      parameters:
        - name: return_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Return transaction was deleted successfully
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid return id"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '404':
          description: Return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Return not found"
                name: "NotFoundError"
        '420':
          description: Cannot delete a Reimbursed return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 420
                message: "Cannot delete a reimbursed return"
                name: "InvalidStateError"
  /returns/sale/{sale_id}:
    get:
      tags:
        - returns
      summary: Get returns for a specific sale
      description: Retrieves all return transactions associated with a sale. 
        Requires **Administrator**, **ShopManager**, or **Cashier** role.
      operationId: getReturnsBySale
      security:
        - bearerAuth: []
      parameters:
        - name: sale_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of returns for the sale
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Return'
        '400':
          description: Invalid or missing id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid return id"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
  /returns/{return_id}/items:
    post:
      tags:
        - returns
      summary: Add a product to a return transaction
      description: |
        Adds a product and quantity to a return transaction.
        Returned quantity must be positive and cannot exceed the quantity sold.
        Requires **Administrator**, **ShopManager**, or **Cashier** role.
      operationId: addProductToReturn
      security:
        - bearerAuth: []
      parameters:
        - name: return_id
          in: path
          required: true
          schema:
            type: integer
        - name: barcode
          in: query
          required: true
          schema:
            type: string
        - name: amount
          in: query
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Product successfully added to return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid return id"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '404':
          description: Return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Return not found"
                name: "NotFoundError"
        '420':
          description: Invalid sale state for return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 420
                message: "Cannot modify a closed return"
                name: "InvalidStateError"
    delete:
      tags:
        - returns
      summary: Remove a product from a return transaction
      description: Decreases or removes a product line from a return transaction. 
        Requires **Administrator**, **ShopManager**, or **Cashier** role.
      operationId: removeProductFromReturn
      security:
        - bearerAuth: []
      parameters:
        - name: return_id
          in: path
          required: true
          schema:
            type: integer
        - name: barcode
          in: query
          required: true
          schema:
            type: string
        - name: amount
          in: query
          required: true
          schema:
            type: integer
      responses:
        '202':
          description: Product successfully removed or decreased
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid return id"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '404':
          description: Return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Return not found"
                name: "NotFoundError"
        '420':
          description: Invalid sale state for return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 420
                message: "Cannot remove items from a closed return"
                name: "InvalidStateError"
  /returns/{return_id}/close:
    patch:
      tags:
        - returns
      summary: Close a return transaction
      description: |
        Closes an **OPEN** return transaction, setting it to **CLOSED** and restores product quantities to inventory.
        **Whenever is invoked on an EMPTY OPEN return (so, just created with 0 products inside) the return has to be deleted from the database.** 
        Requires **Administrator**, **ShopManager**, or **Cashier** role. 
      operationId: closeReturn
      security:
        - bearerAuth: []
      parameters:
        - name: return_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Return closed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid return id"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '404':
          description: Return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Return not found"
                name: "NotFoundError"
        '420':
          description: Invalid sale state for return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 420
                message: "Invalid Return state to be closed"
                name: "InvalidStateError"
  /returns/{return_id}/reimburse:
    patch:
      tags:
        - returns
      summary: Reimburse a return transaction
      description: |
        Marks a **CLOSED** return as **REIMBURSED** and updates system balance. 
        Requires **Administrator** or **ShopManager** role.
      operationId: reimburseReturn
      security:
        - bearerAuth: []
      parameters:
        - name: return_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Return successfully reimbursed
          content:
            application/json:
              schema:
                type: object
                properties:
                  refund_amount:
                    type: number
                    format: float
                    example: 25.5
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid return id"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '404':
          description: Return not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Return not found"
                name: "NotFoundError"
        '420':
          description: Invalid sale state for return
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 420
                message: "Return must be closed before reimbursement"
                name: "InvalidStateError"
...